// data/github/current/ui/packages/pull-request-viewer/components/header/__generated__/PullRequestHeaderWrapperQuery.graphql.ts
export type PullRequestHeaderWrapperQuery$variables = {
  endOid?: string | null | undefined;
  number: number;
  owner: string;
  repo: string;
  singleCommitOid?: string | null | undefined;
  startOid?: string | null | undefined;
};
export type PullRequestHeaderWrapperQuery$data = {
  readonly repository: {
    readonly pullRequest: {
      readonly id: string;
    } | null | undefined;
    readonly " $fragmentSpreads": FragmentRefs<"PullRequestHeaderWrapper_repository">;
  } | null | undefined;
  readonly viewer: {
    readonly " $fragmentSpreads": FragmentRefs<"PullRequestHeaderWrapper_user">;
  };
};
export type PullRequestHeaderWrapperQuery = {
  response: PullRequestHeaderWrapperQuery$data;
  variables: PullRequestHeaderWrapperQuery$variables;
};